Add tests for mutex util module

+++ lib/vagrant/util/file_mutex.rb
@@ -22,7 +22,7 @@ module Vagrant
       end
 
       def unlock
-        File.delete(@mutex_path)
+        File.delete(@mutex_path) if File.file?(@mutex_path)
       end
     end
   end
